fix errors and address comments
