+++ packages/astro-language/src/language.ts
@@ -1,14 +1,11 @@
 import { parse } from "@astrojs/compiler/sync";
-import type { DiagnosticMessage, RootNode } from "@astrojs/compiler/types";
+import type { RootNode } from "@astrojs/compiler/types";
 import { getLanguagePlugin } from "@astrojs/ts-plugin/dist/language.js";
-import {
-	getPositionOfColumnAndLine,
-	type LanguageReport,
-	type SourceFileWithLineMap,
-} from "@flint.fyi/core";
+import type { SourceFileWithLineMap } from "@flint.fyi/core";
 import { setTSExtraSupportedExtensions } from "@flint.fyi/ts-patch";
 import { createVolarBasedLanguage } from "@flint.fyi/volar-language";
 
+import { astroCompilerDiagnosticToLanguageReport } from "./astroCompilerDiagnosticToLanguageReport.ts";
 import { extractDirectives } from "./extractDirectives.ts";
 
 setTSExtraSupportedExtensions([".astro"]);
@@ -16,23 +13,6 @@ setTSExtraSupportedExtensions([".astro"]);
 export interface AstroServices {
 	astro: {
 		ast: RootNode;
-	};
-}
-
-export function astroCompilerDiagnosticToLanguageReport(
-	fileName: string,
-	source: SourceFileWithLineMap,
-	diagnostic: DiagnosticMessage,
-): LanguageReport {
-	// location.line and location.column are both 1-indexed
-	const begin = getPositionOfColumnAndLine(source, {
-		column: diagnostic.location.column - 1,
-		line: diagnostic.location.line - 1,
-	});
-	return {
-		code: `ASTRO${diagnostic.code}`,
-		range: { begin, end: begin + diagnostic.location.length },
-		text: `${fileName}:${diagnostic.location.line}:${diagnostic.location.column} - ${diagnostic.text}${diagnostic.hint ? ` (${diagnostic.hint})` : ""}`,
 	};
 }
